requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==1.4.1
Pillow==9.4.0
torch==2.3.0
torchmetrics==0.8.0
torchvision==0.18.0
tqdm==4.67.0
